reorganise separate clients into their own directory
This PR implements #250 by moving infrastructure clients into app/clients/ directory, keeping only business logic in services/. No architectural or behavioral changes.
