+++ app/clients/client_base.py
@@ -13,7 +13,7 @@ from app.config import Prompt, settings
 from app.database.db import create_new_message_by_fields, get_user_message_history
 from app.database.enums import MessageRole
 from app.database.models import Message, User
-from app.services.whatsapp_service import whatsapp_client
+from app.clients.whatsapp_client import whatsapp_client
 from app.tools.tool_manager import ToolManager
 from app.utils.message_processor import MessageProcessor
 from app.utils.prompt_manager import prompt_manager
