+++ app/services/onboarding_service.py
@@ -4,7 +4,7 @@ from app.database import db
 from app.database.enums import MessageRole, OnboardingState
 from app.database.models import User
 from app.services.flows.flow_service import flow_client
-from app.services.whatsapp_service import whatsapp_client
+from app.clients.whatsapp_client import whatsapp_client
 from app.utils.string_manager import StringCategory, strings
